+++ Assets/Enemy.prefab
@@ -17,6 +17,7 @@ GameObject:
   - component: {fileID: 5007733490338877783}
   - component: {fileID: 8074301482262053757}
   - component: {fileID: 6689704014182575917}
+  - component: {fileID: 1213224063824967014}
   m_Layer: 0
   m_Name: Enemy
   m_TagString: Untagged
@@ -173,3 +174,25 @@ MonoBehaviour:
   target: {fileID: 0}
   enemyHealth: {fileID: 8074301482262053757}
   attackDamage: 10
+--- !u!195 &1213224063824967014
+NavMeshAgent:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2287028595195033631}
+  m_Enabled: 1
+  m_AgentTypeID: 0
+  m_Radius: 0.5
+  m_Speed: 6
+  m_Acceleration: 8
+  avoidancePriority: 50
+  m_AngularSpeed: 120
+  m_StoppingDistance: 0
+  m_AutoTraverseOffMeshLink: 1
+  m_AutoBraking: 1
+  m_AutoRepath: 1
+  m_Height: 1
+  m_BaseOffset: 0.5
+  m_WalkableMask: 4294967295
+  m_ObstacleAvoidanceType: 4
