+++ Assets/Enemy2.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d7fe586a40573d945b824ef8a35bb7dc
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
