+++ Assets/EnemyAC.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3647a072a7ed2d84d980a55b906bb926
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
