+++ Assets/Model.obj
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:a4bd1bd3ca40ef3e75d9e353259b5ab1c120cf47a5b8a280d5fde7d6baf6a6f2
+size 1005445
