+++ Assets/Scenes/Dungeon Defenders/NavMesh.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 295d6ea839b121d48a4c9d0f787af334
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
