+++ Assets/Scripts/EnemyScripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 80aa27d8889295e4982dc5f62b893bb5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
