+++ Assets/Scripts/EnemyScripts/EnemySwordAttack.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 93a8f703e4f050444bf0dbb01db7a06d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
