+++ Assets/Scripts/PlayerScripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 760f602b3bc6107478e49397723f7ea7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
