+++ Assets/Scripts/PlayerScripts/CameraController.cs.meta
